+++ src-app/lib/util/ui/settings/settings.js
@@ -92,7 +92,7 @@ class SettingsUI {
             SettingsUI._copyAttr(
                 {type: type, autocomplete: "off"},
                 attr,
-                ["id", "onclick", "oninput", "onchange", "min", "max", "step", "checked", "value", "name", "placeholder", "spellcheck"]
+                ["id", "onclick", "oninput", "onchange", "min", "max", "step", "checked", "value", "name", "placeholder", "spellcheck", "className"]
             ));
     }